ci-seo-agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
chromadb==1.5.5
apscheduler==3.11.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-analytics-data==0.20.0
mysql-connector-python==9.1.0
anthropic==0.85.0
openai==2.29.0
python-dotenv==1.0.1
requests==2.32.5
numpy==2.4.3
